Add type casts to printf
This causes warnings in the civetweb CodeQL build: https://github.com/civetweb/civetweb/security/code-scanning/2?query=ref%3Arefs%2Fheads%2Fmaster

+++ src/check_error.c
@@ -61,7 +61,7 @@ void *emalloc(size_t n)
 
     p = malloc(n);
     if(p == NULL)
-        eprintf("malloc of %u bytes failed:", __FILE__, __LINE__ - 2, n);
+        eprintf("malloc of %lu bytes failed:", __FILE__, (int)__LINE__ - 2, (unsigned long)n);
     return p;
 }
 
@@ -71,6 +71,6 @@ void *erealloc(void *ptr, size_t n)
 
     p = realloc(ptr, n);
     if(p == NULL)
-        eprintf("realloc of %u bytes failed:", __FILE__, __LINE__ - 2, n);
+        eprintf("realloc of %lu bytes failed:", __FILE__, (int)__LINE__ - 2, (unsigned long)n);
     return p;
 }
